Maint: Don't load facts twice during specs
Previously, the spec_helper was prepending facter's lib directory to the
$LOAD_PATH, which had the side-effect of causing the zfs and zpool commands
to be executed 5 times, not 3 (as specified in the comments).

Originally, this was noticed because we were having to add `.at_least_once`
to expectations even though the method should have only been called once.

This commit changes the spec_helper to no longer prepend facter's top-level
lib directory to the LOAD_PATH. In doing so, it causes the virtual fact
resolutions to be evaluated in a different order than the virtual tests
expected. In particular, the tests we're trying to verify the more-specific
resolution, but were actually runnig with the more generic one.

This commit adds a higher weight to the more specific virtual facts,
since that's the expected behavior, and it ensures the tests reliably
test the resolution they trying to set their expectations on.

Paired-with: Patrick Carlisle <patrick@puppetlabs.com>