(#16668) Use WMI to collect IPv4 address
Previously, we were calling `IPSocket.getaddress(Socket.gethostname)` to
resolve the `ipaddress` fact on Windows, but using WMI-based methods
to resolve other IP related facts, e.g. interfaces. This led to inconsistent
information -- ipaddress from one interface, netmask from another.

This commit introduces a `get_preferred_network_adapters` method that
returns an ordered list of WIN32OLE objects that wrap
Win32_NetworkAdapterConfiguration objects. The list is ordered by the
IPConnectionMetric. If two interfaces have the same metric, e.g. host with
multiple NICs, then the order is based on the adapter bindings[1].

It is possible for a network adapter configuration to contain IPv4 and/or
IPv6 addresses, but only the former will be considered valid for the
ipaddress fact.

[1] http://support.microsoft.com/kb/894564