(#16668) Use WMI to collect macaddress fact
Previously, facter would report on the first macaddress associated
with any interface that had IP enabled. This could easily be an
interface with a high cost metric, or an IPv6 only interface.

This commit refactors the macaddress fact to use the same ordered list
of network adapters as is done for ipaddress, to ensure we return a
consistent set of facts.