(#16668) Use WMI to collect IPv6 address
Previously, we were using netsh.exe to collect the ipaddress6 fact, but it
relied on en_US output that wasn't internationalizable.

This commit uses WMI, like the ipaddress fact, and handles issues like
multiple NICs, both IPv4 and IPv6 IP addresses on a single NIC,
per-interface metrics, and the binding order of interfaces.