(#21954) Add support for Mageia OS to osfamily.rb
The addition of Mageia OS in
https://github.com/puppetlabs/facter/commit/e2f040a5bc4ed7539ecc54a5c3d26f93a9ed6a18
(Ticket #9929) was incomplete.
Mageia is derived from Mandrake and Mandriva, and so is in the same family.
This patch adds support to return osfamily='Mandrake' on a Mageia OS system,
rather than the generic 'Linux' kernel value, as well as adjusting the tests
expecting the kernel fact for the Mageia operating system.