Maint: Simplify how interfaces are enumerated
Previously, we were enumerating each WIN32OLE object, only to have the
caller append the yielded object to an array and then call Enumerable
methods on the array.

This commit simplifies the method to return an Array of WIN32OLE objects,
so the caller can call Enumerable methods directly on it.