(#21738) Add .cmd / Update PowerShell parser
This updates the PowerShell parser to add a switch for the Bypass execution
policy which means that the execution policy will not need to be set
prior to running facter. This also adds some switches (NoProfile,
NoLogo, NonInteractive) that speed up the execution of external powershell
facts.

This also adds the extension .cmd as a valid file extension as part of the
ScriptParser. This functions in the same way as .bat.