(#21738) Add .cmd / Update powershell parser
This adds a call to the bypass execution policy which means someone will not
need to set the execution policy prior to running facter. This also adds
some switches (NoProfile, NoLogo, NonInteractive) that speed up the
execution of external powershell facts.

This also adds the extension .cmd as a valid file extension as part of the
ScriptParser. This functions in the same way as .bat.