Maint: Don't output "No facts loaded from" during tests
Previously, running `bundle exec rspec spec/unit/memory_spec.rb` would
output the warning "No facts loaded from ...". The reason was because
the memorytotal fact is defined in a file whose name doesn't match,
and the code that forces facter to load the memory facts was within an
inner describe block.

See #7670 and commit 2255abee7b for more information about facter's
quirky loading behavior.

This commit just moves the `load(:memory)` before each to the
outermost describe block. This ensures that all of the memory-derived
facts are "preloaded" so facter doesn't print a warning that it can't
find them.