(#22322) Squash stderr on Solaris and AIX when there is no swap
This patch redirects stderr to /dev/null on SOlaris and AIX when the swap -l command is issued. If no swap device is present, then the message "No swap devices configured" is printed to stderr. This, in addition to causing unnecassary log messages, causes certain cron jobs used by Puppet Enterprise to generate emails to root.