(#22233) Conform to code standards
Without this patch the documentation for
Facter::Util::GCE.with_metadata_server does not match the
implementation.  The implementation appears to have been copied from
Facter::Util::EC2.with_metadata_server which does take a block as an
argument.  The GCE implementation does not take a block but the
documentation mentions it does.

This is a problem because we have two methods of the same name that have
different designs, which is confusing.

This patch addresses the problem by modifying the GCE implementation of
with_metadata_server to accept a block.  The reason the block was
originally removed was likely because all of the metadata necessary to
define the facts may be gathered with a single call.  Unlike EC2,
there's no need to first check if the service is responsive then call
back out to the service to get the response.  In GCE if the service is
responsive then the initial response is all that's required.

This patch augments the original design by requiring the block passed to
with_metadata_server to accept a single string argument.  The method
will pass the body of the initial response to the block, which will
usually be JSON, allowing the block to define the fact values based on
the initial response rather than requiring a subsequent request.