Maint: Only try to install json gem for ruby 1.8
The json gem contains native code, so a compiler must be present in
order to install the gem (bundle install). On Windows, we intentionally
do not install the devkit, mainly for security reasons.

However, the json gem is only required when running ruby 1.8, so we
confine the Gemfile accordingly. This resolves the issue on Windows,
because we only support ruby 1.9 on that platform.