(#12504) Read non-configurable file for operatingsystemrelease on Ubuntu
Without this patch, the operatingsystemrelease for Ubuntu is gathered by
parsing /etc/issue. This is a problem as /etc/issue is meant to have
configurable content to optionally be displayed when logging in to a
system. This patch uses /etc/lsb-release which is meant to have static
data describing the release version.