(#22619) Error when NetConnectionId is missing
When someone is using something like failover clustering, an adapter
configuration is created and set to IPEnabled but the underlying network
adapter is set to NetEnabled=False. Further, the NetConnectionId is empty and
causes facter to error. This adds a check to remove items where the adapter
itself is NetEnabled=False as well. This also adds a check to not add empty or
nil items to the list of interfaces.