(#9546) Do not execute com, cmd, exe, or bat files if not on windows
Facter currently only executes com, exe, and bat files if on a windows
system.  It should also exclude these on unix flavored systems since
the interpreter for these "cmd.exe" will not be present.

This also supports the use of pluginsync for external facts where in a
hybrid environment external facts for windows may end up getting syncd
to a unix system.

Conflicts:
	spec/unit/util/parser_spec.rb