Align hostname resolution to domain resolution
There are situations under Linux where `hostname` may return a different short name than the short name appearing in `hostname -f`. This affects the output of factor's version of FQDN which is a join of hostname.rb and domain.rb. The kernel receives it's FQDN via hostname -f in linux and as such the hostname resolution should reflect this.