(FACT-649) Cherry-pick acceptance setup fix for RedHat / Fedora back into master
This commit cherry-picks work done on the install_utils acceptance setup
back into master, which was lost during the facter-2 / master merge.

Original commit (eeb4065):
(maint) Recognize `fc` as a type of redhat platform in acceptance setup

The beaker config file for fedora 18 in this project has a platform
starting with `fc` instead of `fedora`.