(Maint) Execute test on all agents using beaker's DSL method
Previously, the test was only executing on the first agent host, and it
attempted to run facter directly, which fails on windows when executing
via ssh, because facter is a .bat file, not an executable.

This commit updates the test to run on all agent hosts, and it uses
beaker's built-in DSL method for executing facter, which ultimately
executes:

    cmd /c facter --puppet test_fact1

It also uses beaker's method for passing environment variables to the
command being executed.