(#22944) Reload external facts after the collection is flushed
Previously, external facts were not reloaded after
`Facter::Util::Collection#flush` was called, which doesn't match the
semantics of the `flush` method.

This commit ensures we are capable of reloading external facts after
`flush` is called.

Note that this isn't an issue when `Facter.clear` is called, because
that discards the collection entirely, so external facts will be
reloaded if necessary.