(#23377) Fix parsing of 'zfs -?' stdout to allow zfs version detection
Facter detects zfs version by first running 'zfs -?' to verify that it
can call 'zfs upgrade -v' safely. Prior to this patch, the parsing
of the output of 'zfs -?' assumed a different (earlier?) format for the
output of 'zfs -?' and thus would miss that the 'zfs upgrade' sub-command
is in fact supported.

This patch loosens the regex so that 'upgrade' doesn't have to be the
first word on a line, and also adds a new fixture to match the
current output of 'zfs -?'.