(fact-79) Commonize non-nil physicalprocessorcount values to string
Prior to this commit, physicalprocessorcount returned an integer on
non-sunos platforms, but a string on sunos. Also, all the other
processorcount-related facts returned strings regardless.

For consistency cross-platform and cross-processorcount-facts,
this changes all physicalprocessorcount values to be strings.

Note this doesn't change the one case where physicalprocessorcount
returns nil, as that is being addressed by fact-107.