(FACT-65) Collections should be able to directly create facts
The existing `Collection#add` method would create a new fact and
resolution at the same time. This means that facts themselves could not
be created and manipulated. This commit adds `#define_fact` so that new
facts can be created by themselves.

This commit also removes tests inside of the collection for the
:ldapname fact option since that's going the way of the dinosaurs.