(FACT-134) Normalize copies of fact string values
Previously, facter would fail to normalize a frozen, non-UTF8 encoded
string, such as the value returned by the `rubyversion` fact.

This commit changes the normalization module to make a copy of the
resolved value, and perform string encoding on that. It also undoes the
change to Factet::Util::Loader that dup'ed environment variables of the
form `FACTER_*`.

Paired-with: Andy Parker <andy@puppetlabs.com>