(FACT-239) Expose resolution type controls
This commit exposes the ability to create different types of
resolutions. When a resolution is being created the `:type` option can
be specified, creating either a simple resolution (using `:type =>
:simple`) or aggregate (using `:type => :aggregate`). If no type is
given the type defaults to simple.

If a resolution is redefined the requested type is compared against the
currently existing resolution. If the types are mismatched an error will
be raised since there's no particularly good way to resolve this
discrepancy.