(maint) Remove Facter.method_missing
The ability to use method_missing was added in 2006 but was never
documented, explained, or generally acknowledge. It was a constant
surprise when invoking a non-existent method on Facter would cause all
facts to be loaded and possibly a completely nonsensical value would be
returned. This commit removes the given code so we can return to a more
sane state of affairs.