(maint) Don't warn on non-existent paths
Facter was emitting warnings whenever a path specified by
`Facter.search_path` was not a directory, but Puppet was automatically
adding two paths via the `factpath` setting that seem to never exist.
This could cause hundreds of warnings to be generated, none of which are
really that useful. This commit changes the warnings to only be emitted
when a manually specified path is not absolute and silently ignores all
paths that aren't directories.