(maint) Reset facter search path when resetting facts
The previous implementation of Facter search paths was only additive and
there was no way to remove Facter search paths once added. This is
especially problematic when running the Puppet specs which wind up
reloading Facter frequently and adding search paths, and can wind up
adding thousands of meaningless paths. This commit resets the Facter
search path on `Facter.reset` so that when facts are reset, the known
search pats are reset as well.