(FACT-273) Resolutions should be able to reference their facts
The fact-resolution relationship is one to many, and there is
information in the fact that we need in the resolution. This commit
directly couples facts and resolutions for this purpose.

This commit also adds a healthy refactoring of the fact and resolution
tests to use modern rspec expectations and use imperative phrasing in
specifications.