(FACT-322) Remove special casing of resolution command
Resolutions could accept either a command to execute or a block to run
to evaluate when resolving. However this behavior meant that special
behavior had to be inserted into the execution command in case a command
returned nil. This commit converts the #setcode(String) invocation to
create a block that will evaluate a command and normalize the output so
that Facter::Core::Execution doesn't have to be responsible for this.