(FACT-332) Add cfpropertylist task to facter
As cfpropertylist is no longer vendored within facter, but it is still needed
at runtime for facter, this commit adds a simple task to fetch, unpack, and
stage cfpropertylist within the facter libdir, where it will be installed
automatically in the apple package. This task is intended to be used only when
building apple packages.