(maint) Load extra tasks before loading packaging
Because the apple task now depends on the cfpropertylist task, the packaging
will verify that such a task exists when loaded. This means that the extra
facter tasks must be loaded before we load the packaging tools. This commit
moves the facter task loading to before the packaging tool loading.