(maint) Update docs to reflect use of facter-2 branch
As we prepared for the release of Facter 2.0, we decided
to do that development on a facter-2 branch rather than
master. For details on this decision, see:

https://groups.google.com/forum/#!msg/puppet-dev/Q24GLe6s1_4/hPwHgvVhnpIJ

This commit reflects this change into the COMMITTERS.md and
CONTRIBUTING.md documents. It's mostly replacing 'master' with
'facter-2' in the appropriate places, but it also removes some
old assumptions, e.g. about per-release branches.