Correct assumed regex for AIX netmask fact
On AIX, the printed output of ifconfig is similar to the best family with
respect to netmask (that is, prepends the value with 0x):

[0] [AIX] root@pe-aix-61-agent:~ # uname -a
AIX pe-aix-61-agent 1 6 00F7FD3A4C00 powerpc unknown AIX
[0] [AIX] root@pe-aix-61-agent:~ # /etc/ifconfig en0
en0: flags=1e080863,480<UP,BROADCAST,NOTRAILERS,RUNNING,SIMPLEX,MULTICAST,GROUPRT,64BIT,CHECKSUM_OFFLOAD(ACTIVE),CHAIN>
        inet <redacted> netmask 0xffffff00 broadcast <redacted>
        tcp_sendspace 262144 tcp_recvspace 262144 rfc1323 1

Note netmask 0xffffff00, which is 255.255.255.0.

This means the regex block in facter/util/netmask.rb that AIX belongs to is the
one with the various bsd variants.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>