Add better network fact support on AIX
This commit updates network/ip fact support on AIX to include the following facts:

* interfaces
* interface-specific IP addresses (e.g. ipaddress_en0)
* netmask addresses (e.g. netmask_en0)
* network-addresses (e.g. network_en0)

It does so by leveraging the existing Facter::Util::IP::Base class mostly. This
class is totally sufficient for the majority of network facts on AIX. I have
added fixtures and unit tests, though I admit my knowledge of the facter
spec+unit+fixture layout is not great, so forgive me if this is done in an
outdated or incorrect way. I'd be glad to fix up where needed.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>