Pin ffi to 1.9.0 to match puppet
Previously I specified the ffi gem with a pessimistic version,
but that conflicted with puppet's fixed pinning of ffi to 1.9.0.
Puppet's fixed pinning is really just for windows (the only platform
that requires ffi currently), but it would take a little more work
than I am up for to puppet/Gemfile to support platform dependencies
more loosely, so taking the easy out and just pinning it here to
match. For now.