(maint) Return nil on dns command lookup fail
This updates the domain fact to ignore the output of nonexistent
commands, rather than raising an error. While it would be better to
refactor the domain fact into smaller resolutions that can be confined
based on the presence of the command that's a more ambitious refactor
than what I can do right now.