(FACT-375) Include facter schema for acceptance tests
In order to run acceptance tests the acceptance/ and schema/ directories
need to be included in acceptance-artifacts.tar.gz. Originally we were
cding to acceptance/ and tarring up everything in that directory which
made it difficult to include another top level directory. This commit
changes the logic to tar up the acceptance/ and schema/ directories from
the project root so that we can easily preserve the original directory
structure.