(FACT-185) Provide flattened version of EC2 facts
To maintain compatibility with the existing EC2 facts, we need to return
both a structured version of the EC2 facts as well as flattened
versions. This has the unfortunate side effects of forcing fact
evaluation at fact load time, but between the options of managing a
singleton object, duplicating the API queries, and forcing the
evaluation, the last one seems the least bad.