(maint) Update to build out-of-source.
Updating CMake files to support building "out-of-source".  This is where
you create a directory specifically for containing the build files and
generating the Makefile in that directory.  In addition to keeping the
source directories from being cluttered with build outputs, this also
allows us to generate different CMake configurations (debug vs. release,
for example) side-by-side.

Moving the rapidjson source out of the lib directory and into the vendor
directory.