(FACT-460) Modify $GEM_SOURCE for CI bootstrap
It is preferable to override the authoritative source (rubygems.org)
with our own internal resources, as opposed to the other way around.
This change ensures that if $GEM_SOURCE is set in the environment,
it is honored. Otherwise, we fall back to the public authority,
rubygems.org.

This change is analogous to the following line in our Gemfiles:

    source ENV['GEM_SOURCE'] || "https://rubygems.org"