Add docker detection to virtual and is_virtual facts
Without this patch facter is not detecting the Docker virtual
environment.  This is a problem because docker containers are a
lightweight virtualization approach and as such should be flagged.

This patch addresses the problem by specifically looking for the docker
specific substring in the progress group hierarchy of the init process.
This is very similar, but distinct, from the LXC detection.

Docker builds upon generic Linux containers and customizes the detection
to indicate a docker specific environment.  This patch detects the
docker specific environment.