(maint) Fix Linux physicalprocessorcount fact to handle missing topology information.
CPU topology information may not be present in /proc/cpuinfo.  If it is
missing, we were not properly counting the physical CPUs present on the system.

The fix is to first attempt to use sysfs to count CPUs as it more
reliable for the topology information.  If sysfs isn't present, we fall
back to using topology information in /proc/cpuinfo.

This change matches what we do in Ruby facter.