(FACT-451) Only use /etc/os-release on Cumulus Linux
Without this commit the Cumulus Linux operatingsystem check always uses
the contents of /etc/os-release to set the operatingsystem fact. This is
backwards incompatible as this field may not resolve to the same value
as the older detection. This commit adds a Cumulus Linux specific
resolver that checks /etc/os-release early so it resolves before Debian,
but only returns a value if the operatingsystem is Linux.