(CFACT-19) Implement cfacter gem.
Re-implementing the cfacter gem.
Removing JSON dependency on the gem since we now use callbacks to
populate a hash instead of emitting and parsing a JSON string.
projects.
Adding Rakefile to gem directory with tasks for packaging and installing
the gem.
Adding tasks for running specs.
Adding specs that do some basic validation of the cfacter gem.
Updating README for instructions on how to build, test, and install the
gem.