(maint) Fix domain and fqdn facts when domain resolves to hostname.
In the case where the domain is the same as the hostname, treat it like the domain could not be resolved.
This prevents us from setting the domain and fqdn facts to the same value as the hostname.
Ruby facter returns nil for domain and fqdn facts under this condition.