(fact-233) Pass '-f all' to nmcli to ensure the dhcp_server field is available
Prior to this change 'nmcli d' was invoked without specifying which
fields were needed. This happened to work on fedora 19 and fedora 20,
but doesn't work on redhat 7, which defaults to return a restricted
set of fields for devices.

This patch specifies '-f all' to nmcli so that all fields are returned.
I've tested this change on all 3 of the above distros. It works on all 3
and doesn't seem notably slower (I had some concern that maybe this would
cause a slowdown, so wanted to check that out.)