(maint) Set RPATH for non-default lib install path.
If we're using an install prefix that is not targeting a directory
normally searched by the loader, tell CMake to set an install RPATH to the
directory containing libfacter.so.

Also instructing CMake to add "@rpath/" to the install name of libfacter
when building and installing to a non-default lib directory on OSX.  This
allows libfacter to be loaded from the RPATH for that platform.