(FACT-233) Fix nmcli version detection for DHCP server fact.
Previously, we were doing a simple "to_i" on the extracted version
string from nmcli to determine which command line syntax to use.  If the
version was greater than 0.9.9.0, we need to use the "show" command
instead of the "list" command.  This was done by checking the result
against 990 ("0990".to_i, meaning version "0.9.9.0").

However, this won't work with version strings such as "0.9.8.10" since
"09810".to_i > 990.

The fix is to extract the version components of the string and
explicitly check against version 0.9.9.0 or higher.