(FACT-547) Limit the output of prtdiag on Solaris
prtdiag is used to determine manufacturer and productname on Solaris.

If you run it on a big box like a SPARC T5 it produces a lot of output,
and might take some time, this produces timeouts leading to missing facts.

The simple addition of a pipe to head avoid prtdiag to produce it full output.

The first lines should always contain the product information.