(FACT-550) Skip macosx and posix specific tests on windows
Commit 5b90132ee unvendored CFPropertyList, and we expressed the
CFPropertyList gem dependency as:

    gem 'CFPropertyList', :platform => ruby

However, bundler interprets the `ruby` platform as "C Ruby (MRI) or
Rubinius, but NOT Windows"[1]. This means the gem is installed everywhere,
but Windows.

This causes the macos_spec to fail on Windows, but for some reason Jenkins
CI thinks the job passes.

This commit skips testing macos_spec on windows, and skips posix_spec which
fails due to ':' not being the PATH separator on windows.